tests with qbittorrent as client and server running in same time passed successfully

+++ resync.py
@@ -1,23 +1,31 @@
-import os, re, linecache, time, requests, json, html
+import json
+import os
+import re
+import requests
+import time
+import yaml
 
 
 def getFromCategory(idCat, CFcookies):
-    url = "https://www4.yggtorrent.li/rss?action=generate&type=subcat&id=" + idCat + "&passkey=TNdVQssYfP3GTDnB3ijgE37c8MVvkASH"
+    url = "https://www4.yggtorrent.li/rss?action=generate&type=subcat&id=" + idCat + \
+          "&passkey=TNdVQssYfP3GTDnB3ijgE37c8MVvkASH"
     print(url)
     headers = {'User-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)'}
     return (requests.get(url, cookies=CFcookies, headers=headers)).text
 
 
-def ManageTorrents(rssData, CFcookies, idCat):
+def ManageTorrents(rssData, CFcookies, idCat, categories):
     # extract from rss feed all id torrent
     rssTorrentsListId = re.findall("id=[0-9]{6}", rssData)
 
-    if os.path.isfile("rss/" + idCat + ".xml"):
+    # check if rss file for category requested is available
+    if os.path.isfile("rss/" + idCat + ".xml") and (int(idCat) not in categories):
         oldRssFile = open("rss/" + idCat + ".xml", "r")
-        oldRssString = oldRssFile.read(); oldRssFile.close()
+        oldRssString = oldRssFile.read()
+        oldRssFile.close()
         oldRssTorrentsListId = re.findall("id=[0-9]{6}", oldRssString)
         for oldIdTorrent in oldRssTorrentsListId:
-            if oldIdTorrent not in rssTorrentsListId and os.path.isfile("torrents/" + (re.split("=", oldIdTorrent)[1]) + ".torrent"):
+            if (oldIdTorrent not in rssTorrentsListId) and (os.path.isfile("torrents/" + (re.split("=", oldIdTorrent)[1]) + ".torrent")):
                 print("removing --> " + (re.split("=", oldIdTorrent)[1]))
                 os.remove("torrents/" + (re.split("=", oldIdTorrent)[1]) + ".torrent")
 
@@ -38,6 +46,7 @@ def ManageTorrents(rssData, CFcookies, idCat):
 
 
 def initSession(id, serverPath):
+    print(serverPath)
     # initialize session identified by id on serverPath flaresolverr
     payload = json.dumps({
         "cmd": "sessions.create",
@@ -49,26 +58,10 @@ def initSession(id, serverPath):
         'Content-Type': 'application/json'
     }
     # TODO : look at response from request to flaresolverr (200/XXX)
-    requests.request("POST", "http://" + serverPath + "/v1", headers=headers, data=payload)
+    requests.request("POST", serverPath + "/v1", headers=headers, data=payload)
 
 
-def getCatRequested():
-    # return categories for which node had been setup
-    fileCat = open("idCat.txt", "r")
-    analyseCategories = fileCat.readlines()
-    catList = []
-    for idCat in analyseCategories:
-        if not (int(idCat) < 2139 or int(idCat) == 2146 or int(idCat) > 2187):
-            catList.append(int(idCat))
-    return catList
-
-
-def getServerIpPort():
-    return linecache.getline('annexes.txt', 1).strip() + ":" + linecache.getline('annexes.txt', 2).strip()
-
-
-def getCookies(serverPath, idSession):
-    url = "http://" + serverPath + "/v1"
+def getCookies(url, idSession):
     payload = json.dumps({
         "cmd": "request.get",
         "url": "https://www4.yggtorrent.li",
@@ -80,39 +73,50 @@ def getCookies(serverPath, idSession):
         'Content-Type': 'application/json'
     }
     cookies = dict()
-    for i in json.loads((requests.request("POST", url, headers=headers, data=payload)).text).get('solution').get(
-            'cookies'):
+    print(url, payload)
+    for i in json.loads((requests.request("POST", url + "/v1", headers=headers, data=payload)).text).get(
+            'solution').get('cookies'):
         cookies[i.get('name')] = i.get('value')
     return cookies
 
-def changeDownloadUrl(rssFeed):
-    return (re.sub("https://www4.yggtorrent.li/rss/", "https://79.142.69.160:36820/", rssFeed))
+
+def changeDownloadUrl(rssFeed, serverURL):
+    return re.sub("https://www4.yggtorrent.li/rss/", serverURL + "/", rssFeed)
+
 
 if __name__ == '__main__':
+    confFile = open('annexes.txt', 'r')
+    serverConfiguration = yaml.safe_load(confFile)
+    confFile.close()
+
     # construct string containing ip and port server
-    serverPath = getServerIpPort()
+    FlaresolverrPath = "http://" + str(serverConfiguration["flaresolverr"]["ipAdress"]) + ":" + \
+                       str(serverConfiguration["flaresolverr"]["port"])
+
+    nodeURL = serverConfiguration["node"]["protocol"] + "://" + str(
+        serverConfiguration["node"]["ipAdress"]) + ":" + str(serverConfiguration["node"]["port"])
+
     # initialize session in flaresolverr server
     id_session = "123456789"
-    initSession(id_session, serverPath)
+    initSession(id_session, FlaresolverrPath)
+
     # read category to be syncing on this node.
-    catlist = getCatRequested()
+    catList = serverConfiguration["Categories"]["id"]
+    subCatList = serverConfiguration["sub-Categories"]["id"]
 
     # infinite loop to resync every X seconds
-    #while True:
-    cookies = getCookies(serverPath, id_session)
-    print(cookies)
-    print(catlist)
-    idCat=2183
-#        for idCat in catlist:
-    print(str(idCat))
-    # get rss feed from idCat
-    rssString = getFromCategory(str(idCat), cookies)
-    # download new torrents
-    #ManageTorrents(rssString, cookies, str(idCat))
-    # write rss feed and erasing old xml file
-    rssString = (changeDownloadUrl(rssString))
-    print(rssString)
-    file = open("rss/" + str(idCat) + ".xml", "w")
-    file.write(rssString)
-    file.close()
-    #time.sleep(300)
+    while True:
+        cookies = getCookies(FlaresolverrPath, id_session)
+        for idCat in subCatList + catList:
+            print(str(idCat))
+            # get rss feed from idCat
+            rssString = getFromCategory(str(idCat), cookies)
+            # download new torrents
+            ManageTorrents(rssString, cookies, str(idCat), catList)
+            # write rss feed and erasing old xml file
+            rssString = (changeDownloadUrl(rssString, nodeURL))
+            file = open("rss/" + str(idCat) + ".xml", "w")
+            file.write(rssString)
+            file.close()
+        print("en pause")
+        time.sleep(300)
